feat(socket): add client and server socket functionality
