+++ libzappy_net/Makefile
@@ -34,7 +34,9 @@ HEADERS = $(INC_DIR)/libzappy_net.h \
 # Source files
 SRC_C = $(SRC_DIR)/libzappy_net.c \
         $(SRC_DIR)/zappy_net_core.c \
-        $(SRC_DIR)/zappy_net_socket.c
+        $(SRC_DIR)/zappy_net_socket.c \
+        $(SRC_DIR)/zappy_net_server.c \
+        $(SRC_DIR)/zappy_net_client.c
 SRC_CPP =
 
 # All source files
